requirements.txt
requests
tensorflow
Pillow
